删除重复处理

+++ iShop/src/main/java/com/bizvane/ishop/controller/AreaController.java
@@ -186,20 +186,22 @@ public class AreaController {
             for (int i = 0; i < ids.length; i++) {
                 logger.info("-------------delete--" + Integer.valueOf(ids[i]));
                 Area area = areaService.getAreaById(Integer.valueOf(ids[i]));
-                String area_code = area.getArea_code();
-                String corp_code = area.getCorp_code();
-                List<Store> stores = areaService.getAreaStore(corp_code, area_code);
-                if (stores.size() == 0) {
-                    areaService.delete(Integer.valueOf(ids[i]));
-                    dataBean.setCode(Common.DATABEAN_CODE_SUCCESS);
-                    dataBean.setId(id);
-                    dataBean.setMessage("success");
-                } else {
-                    dataBean.setCode(Common.DATABEAN_CODE_ERROR);
-                    dataBean.setId(id);
-                    dataBean.setMessage("区域" + area_code + "下有所属店铺，请先处理区域下店铺再删除！");
-                    return dataBean.getJsonStr();
+                if (area != null) {
+                    String area_code = area.getArea_code();
+                    String corp_code = area.getCorp_code();
+                    List<Store> stores = areaService.getAreaStore(corp_code, area_code);
+                    if (stores.size() == 0) {
+                        areaService.delete(Integer.valueOf(ids[i]));
+                    } else {
+                        dataBean.setCode(Common.DATABEAN_CODE_ERROR);
+                        dataBean.setId(id);
+                        dataBean.setMessage("区域" + area_code + "下有所属店铺，请先处理区域下店铺再删除！");
+                        return dataBean.getJsonStr();
+                    }
                 }
+                dataBean.setCode(Common.DATABEAN_CODE_SUCCESS);
+                dataBean.setId(id);
+                dataBean.setMessage("success");
             }
         } catch (Exception ex) {
             dataBean.setCode(Common.DATABEAN_CODE_ERROR);
